new file:   assets/js/popover-config.js 	new file:   assets/vendor/bootstrap/js/bootstrap.bundle.min.js 	new file:   assets/vendor/bootstrap/js/bootstrap.bundle.min.js.map 	new file:   assets/vendor/jquery/jquery-3.4.1.min.js 	new file:   assets/vendor/jquery/jquery-3.5.0.min.js 	new file:   assets/vendor/jquery/jquery-3.5.0.min.map 	new file:   assets/vendor/popper/popper.min.js 	new file:   assets/vendor/popper/popper.min.js.map 	modified:   detail_bulan_pembayaran.php 	modified:   include/js.php 	modified:   include/sidebar.php 	modified:   profile.php 	modified:   siswa.php 	modified:   uang_kas.php 	modified:   user.php

+++ siswa.php
@@ -170,7 +170,9 @@
                             </form>
                           </div>
                         </div>
-                        <a data-nama="<?= $ds['nama_siswa']; ?>" class="btn-delete badge badge-danger" href="hapus_siswa.php?id_siswa=<?= $ds['id_siswa']; ?>"><i class="fas fa-fw fa-trash"></i> Hapus</a>
+                        <?php if ($_SESSION['id_jabatan'] == '1'): ?>
+                          <a data-nama="<?= $ds['nama_siswa']; ?>" class="btn-delete badge badge-danger" href="hapus_siswa.php?id_siswa=<?= $ds['id_siswa']; ?>"><i class="fas fa-fw fa-trash"></i> Hapus</a>
+                        <?php endif ?>
                       </td>
                     </tr>
                   <?php endforeach ?>
